fixed decimal prices

+++ components/cart/CartTile.jsx
@@ -49,7 +49,7 @@ export default CartTile = ({item, onPress, select, cartItems, refetch, data}) =>
         <View style={styles.txtContainer}>
             <Text numberOfLines={1} style={styles.productTxt}>{item.cartItem.title}</Text>
             <Text numberOfLines={1} style={styles.supplier}>{item.cartItem.supplier}</Text>
-            <Text numberOfLines={1} style={styles.price}>${(parseInt(item.cartItem.price) * itemQuantity).toFixed(2)}</Text>
+            <Text numberOfLines={1} style={styles.price}>${(parseFloat(item.cartItem.price) * itemQuantity).toFixed(2)}</Text>
         </View>
         <View style={{display: 'flex', flexDirection: 'row', gap: 10}}>
             <TouchableOpacity onPress={() => {
@@ -57,11 +57,11 @@ export default CartTile = ({item, onPress, select, cartItems, refetch, data}) =>
                     decreaseCartItemQuantity(item.cartItem._id, 1); 
                     setItemQuantity(prevQuantity => prevQuantity > 0 ? prevQuantity - 1 : 0);
                     setCartCount(prevCount => prevCount > 0 ? prevCount - 1 : 0)
-                    setSubTotal(prevTotal => (prevTotal -= parseInt(item.cartItem.price)))
+                    setSubTotal(prevTotal => (prevTotal -= parseFloat(item.cartItem.price)))
                 } else {
                     deleteCartItem();
                     setCartCount(prevCount => prevCount > 0 ? prevCount - 1 : 0)
-                    setSubTotal(prevTotal => (prevTotal -= parseInt(item.cartItem.price)))
+                    setSubTotal(prevTotal => (prevTotal -= parseFloat(item.cartItem.price)))
                 }
                 }
                 }>
@@ -74,7 +74,7 @@ export default CartTile = ({item, onPress, select, cartItems, refetch, data}) =>
                 addToCart(item.cartItem._id, 1);
                 setItemQuantity(prevQuantity => prevQuantity + 1);
                 setCartCount(prevCount => prevCount + 1)
-                setSubTotal(prevTotal => (prevTotal += parseInt(item.cartItem.price)))
+                setSubTotal(prevTotal => (prevTotal += parseFloat(item.cartItem.price)))
                 }}>
                 <SimpleLineIcons 
               name='plus'
